+++ src/main/java/dev/willyelton/crystal_tools/gui/component/ClearFilterButton.java
@@ -6,17 +6,17 @@ import net.minecraft.network.chat.Component;
 
 import static dev.willyelton.crystal_tools.gui.CrystalBackpackScreen.TEXTURE;
 
-public class MatchContentsButton extends CrystalToolsButton {
+public class ClearFilterButton extends CrystalToolsButton {
     private static final int BUTTON_SIZE = 12;
     private static final int TEXTURE_Y = 281;
 
-    public MatchContentsButton(int x, int y, OnPress onPress, OnTooltip onToolTip) {
-        super(x, y, BUTTON_SIZE, BUTTON_SIZE, Component.literal("Match Backpack Contents"), onPress, onToolTip);
+    public ClearFilterButton(int x, int y, OnPress onPress, OnTooltip onToolTip) {
+        super(x, y, BUTTON_SIZE, BUTTON_SIZE, Component.literal("Clear Filters"), onPress, onToolTip);
     }
 
     @Override
     protected void blitButton(GuiGraphics guiGraphics, int textureY) {
-        guiGraphics.blit(TEXTURE, getX(), getY(), 60, textureY, BUTTON_SIZE, BUTTON_SIZE, 512, 512);
+        guiGraphics.blit(TEXTURE, getX(), getY(), 48, textureY, BUTTON_SIZE, BUTTON_SIZE, 512, 512);
     }
 
     @Override